pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "python-project-deployment"
version = "0.1.0"
description = "A scaffolding tool for creating new Python packages with best practices"
readme = "README.md"
requires-python = ">=3.13"
license = {text = "MIT"}
authors = [
    {name = "Magic Man", email = "deleterious420+PythonProjectDeployment@gmail.com"},
]
keywords = ["scaffolding", "template", "python", "package", "project"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "click>=8.3.0",
    "jinja2>=3.1.6",
    "pydantic>=2.12.3",
    "pydantic-settings>=2.10.0",
    "python-dotenv>=1.0.0",
    "email-validator>=2.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.4.2",
    "pytest-cov>=7.0.0",
    "pytest-timeout>=2.3.0",
    "pytest-mock>=3.14.0",
    "black>=25.9.0",
    "isort>=7.0.0",
    "mypy>=1.18.2",
    "ruff>=0.14.3",
    "pre-commit>=4.3.0",
    "bandit>=1.8.6",
    "safety>=3.6.2",
    "sphinx>=8.2.3",
    "sphinx-rtd-theme>=3.0.2",
    "sphinx-autodoc-typehints>=3.5.2",
]

[project.scripts]
scaffold-python = "python_project_deployment.cli:main"

[project.urls]
Homepage = "https://github.com/Magic-Man-us/PythonProjectDeployment"
Repository = "https://github.com/Magic-Man-us/PythonProjectDeployment"

[tool.hatch.build.targets.wheel]
packages = ["src/python_project_deployment"]

[tool.pytest.ini_options]
testpaths = ["tests"]
# Coverage reporting is optional - remove --cov flags for faster test runs
addopts = "-v"
# Uncomment below to enable coverage reporting:
# addopts = "--cov=src/python_project_deployment --cov-report=term-missing --cov-report=html"

[tool.black]
line-length = 100
target-version = ['py312']

[tool.ruff]
line-length = 100
target-version = "py312"
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "B", "C4"]
per-file-ignores = {"tests/*" = ["S101", "S108"]}

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.13"
warn_return_any = false  # Relaxed: don't warn on Any returns
warn_unused_configs = true
disallow_untyped_defs = false  # Relaxed: allow untyped functions
mypy_path = "src"
# Strict mode disabled for easier development
# To enable strict type checking, set the above to true

[dependency-groups]
dev = [
    "sphinx>=8.2.3",
    "sphinx-rtd-theme>=3.0.2",
]
